requirements.txt
telebot
datetime
time

# for .env file support
python-dotenv